+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 Change Log
 ==========
 
+2021-03-29
+----------
+
+ * disable humanised keys by default to preserve the input/error mapping
+ * add prohibited/if/unless based on the same named Laravel validation rules
+ * add mimes replacement to work with Symfony file array
+
 2021-02-05
 ----------
 
@@ -12,7 +19,6 @@ Change Log
  * add tests for various use cases
  * add file upload tests
  * replacement rules to work with Symfony file array / UploadedFile
-
 
 2021-01-25
 ----------